+++ changelog.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unversioned
+    - Bugfixes
+	    - Added ability to detect missing timestamps in ADV data
+
 ## Version 1.0.1
 	- Bugfixes:
 		- ADV and TRDI correlation and amplitude 'dir' dimension values